components: move initialization logic into a separate source file
